+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Bug fixes
 
+- [#115][GL-115] - Spectators can still be invited to and join a team
+    - Added `team.message.spectators-cannot-join` to `lang.yml`
 - [#105][GL-105] - The prevent-player-from-leaving-team config option has no effect
 - [#104][GL-104] - NullPointerException when using the pick-random-world-from-list setting
 - [#102][GL-102] - Players are teleported to the wrong location when respawning as spectators after dying during the deathmatch
@@ -11,6 +13,7 @@
 - [#100][GL-100] - ArrayIndexOutOfBoundsException when using the cave-ores-only setting
 - [#99][GL-99] - Items are sometimes missing from the chest in the Timebomb scenario
 
+[GL-115]: https://gitlab.com/uhccore/uhccore/-/issues/115
 [GL-105]: https://gitlab.com/uhccore/uhccore/-/issues/105
 [GL-104]: https://gitlab.com/uhccore/uhccore/-/issues/104
 [GL-102]: https://gitlab.com/uhccore/uhccore/-/issues/102